# Speed up tasks that use `hostsvars`