--- a/library/cartridge_connect_to_membership.py
+++ b/library/cartridge_connect_to_membership.py
@@ -4,7 +4,7 @@ from ansible.module_utils.helpers import Helpers as helpers
 
 argument_spec = {
     'console_sock': {'required': True, 'type': 'str'},
-    'hostvars': {'required': True, 'type': 'dict'},
+    'facts': {'required': True, 'type': 'dict'},
     'play_hosts': {'required': True, 'type': 'list'},
 }
 
@@ -21,12 +21,12 @@ def probe_server(control_console, uri):
 
 def connect_to_membership(params):
     control_console = helpers.get_control_console(params['console_sock'])
-    hostvars = params['hostvars']
+    facts = params['facts']
     play_hosts = params['play_hosts']
 
     changed = False
 
-    for instance_name, instance_vars in hostvars.items():
+    for instance_name, instance_vars in facts.items():
         if helpers.is_expelled(instance_vars) or helpers.is_stateboard(instance_vars):
             continue