--- a/library/cartridge_get_not_expelled_instance.py
+++ b/library/cartridge_get_not_expelled_instance.py
@@ -3,19 +3,19 @@
 from ansible.module_utils.helpers import Helpers as helpers
 
 argument_spec = {
-    'facts': {'required': True, 'type': 'dict'},
+    'module_hostvars': {'required': True, 'type': 'dict'},
     'play_hosts': {'required': True, 'type': 'list'},
 }
 
 
 def get_one_not_expelled_instance(params):
-    facts = params['facts']
+    module_hostvars = params['module_hostvars']
     play_hosts = params['play_hosts']
 
     not_expelled_instance_name = None
 
     for instance_name in play_hosts:
-        instance_vars = facts[instance_name]
+        instance_vars = module_hostvars[instance_name]
         if helpers.is_expelled(instance_vars) or helpers.is_stateboard(instance_vars):
             continue
 
@@ -26,7 +26,7 @@ def get_one_not_expelled_instance(params):
         errmsg = "Not found any instance that is not expelled and is not a stateboard"
         return helpers.ModuleRes(failed=True, msg=errmsg)
 
-    instance_info = facts[not_expelled_instance_name]['instance_info']
+    instance_info = module_hostvars[not_expelled_instance_name]['instance_info']
 
     return helpers.ModuleRes(changed=False, fact={
         'name': not_expelled_instance_name,