--- a/library/cartridge_get_single_instances_for_each_machine.py
+++ b/library/cartridge_get_single_instances_for_each_machine.py
@@ -3,7 +3,7 @@
 from ansible.module_utils.helpers import Helpers as helpers
 
 argument_spec = {
-    'hostvars': {'required': True, 'type': 'dict'},
+    'facts': {'required': True, 'type': 'dict'},
     'play_hosts': {'required': True, 'type': 'list'},
 }
 
@@ -16,14 +16,14 @@ def get_machine_hostname(instance_vars, instance_name):
 
 
 def get_one_not_expelled_instance_for_machine(params):
-    hostvars = params['hostvars']
+    facts = params['facts']
     play_hosts = params['play_hosts']
 
     machine_hostnames = set()
     instance_names = []
 
     for instance_name in play_hosts:
-        instance_vars = hostvars[instance_name]
+        instance_vars = facts[instance_name]
 
         if helpers.is_expelled(instance_vars):
             continue